+++ plone/app/contentrules/tests/dummy.py
@@ -1,24 +1,24 @@
 from OFS.SimpleItem import SimpleItem
-from zope.interface import implements
+from zope.interface import implementer
 from plone.contentrules.rule.interfaces import IRuleElementData
 from zope.component.interfaces import IObjectEvent
 from plone.uuid.interfaces import IAttributeUUID
 
 
+@implementer(IRuleElementData)
 class DummyCondition(SimpleItem):
-    implements(IRuleElementData)
     element = "dummy.condition"
     summary = "Dummy condition"
 
 
+@implementer(IRuleElementData)
 class DummyAction(SimpleItem):
-    implements(IRuleElementData)
     element = "dummy.action"
     summary = "Dummy action"
 
 
+@implementer(IObjectEvent)
 class DummyEvent(object):
-    implements(IObjectEvent)
 
     def __init__(self, object):
         self.object = object
@@ -30,5 +30,6 @@ class DummyRule(object):
         self.__name__ = name
 
 
+@implementer(IAttributeUUID)
 class DummyNonArchetypesContext(object):
-    implements(IAttributeUUID)
+    pass