+++ plone/app/contentrules/tests/test_action_copy.py
@@ -1,4 +1,4 @@
-from zope.interface import implements
+from zope.interface import implementer
 from zope.component import getUtility, getMultiAdapter
 
 from plone.contentrules.engine.interfaces import IRuleStorage
@@ -16,8 +16,8 @@ from plone.app.testing import TEST_USER_ID
 from zope.component.interfaces import IObjectEvent
 
 
+@implementer(IObjectEvent)
 class DummyEvent(object):
-    implements(IObjectEvent)
 
     def __init__(self, object):
         self.object = object